ai-service/requirements.txt
Flask==2.3.3
gunicorn==20.1.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.3.2
nltk==3.9.0
python-dotenv==1.0.0
